Apply fixes suggested by Clippy version `nightly-2023-08-22`.
I'm preparing a toolchain update PR that will supercede #481. I'd prefer to avoid sending a large PR that mixes these style updates with important `unsafe` code, so I'm splitting the update into two PRs.

+++ unittest/src/fake/ambient_light/mod.rs
@@ -33,7 +33,7 @@ impl AmbientLight {
     pub fn set_value(&self, value: u32) {
         if self.busy.get() {
             self.share_ref
-                .schedule_upcall(0, (value as u32, 0, 0))
+                .schedule_upcall(0, (value, 0, 0))
                 .expect("Unable to schedule upcall");
             self.busy.set(false);
         }
